==> notebooks/01_load_and_validate_raw_data.ipynb <==
### Get the Raw Data

First, we want to get our raw data from the web. This can be done using the ##requests## package. The function *download_one_file_of_raw_data* will do just that. We use a type hint in the function definition to help with debugging. This just means we are expecting a Path to be returned to us. We give the function a year and a month that will be input into the url so that any month and year can be downloaded. To get single digit months, we need to use {month:02d} which pads them with a 0. The data will be downloaded into the desired path which we name when the status_code is equal to 200, which means it was a successful request. If not, we create an exception string. We do this by using open(path, 'wb').write(response.content). The 'wb' mode is used for pictures or other non-text files. So we open the path in write binary mode then we write the content of our response in that path.

from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    '''Download the files from the web.'''
    # establish the url we want to get to use {month:02d} to pad single digits with a 0.
    url = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'

    # requests.get() gets us to the given url
    response = requests.get(url)

    # status code 200 means successful request
    if response.status_code == 200:
        # open the path and write the data to it. wb is writing binary data and the requests.content is binary data
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path, 'wb').write(response.content)
        return path
    
    # if the status code was not success...
    else:
        raise Exception(f'{url} is not available.')

# call the function
download_one_file_of_raw_data(year = 2022, month = 1)

### Load the Data

import pandas as pd

# load our data using pd.read_parquet
rides = pd.read_parquet('../data/raw/rides_2022-01.parquet')

# view first 5 rows
rides.head()

For this project, we will only be working with the pickup time and the pick up location. We can use these two columns as the only columns in our dataframe.

# only use pickup time and pick up location
rides = rides[['tpep_pickup_datetime', 'PULocationID']]

# rename columns
rides.rename(columns = {
    'tpep_pickup_datetime' : 'pickup_datetime',
    'PULocationID' : 'pickup_location_id'
}, inplace = True
)

rides.columns

rides.head(n = 10)

### Validate the Data

If we do not validate our data we may be working with bad data and that will always give us bad results. The first thing we can do to validate this data is to make sure that all rides happened in January of 2022 like they are supposed to. To see this, we use **.describe()** and see that the min calue is acutally fro 2008 and the max value is actually in May of 2022 so these rows need to be removed.

# did all rides happen in Jauary 2022?
rides.describe()

# make sure all rows are within our time frame.
# rides must be greater than or equal to January 01, 2022
rides = rides[rides.pickup_datetime >= '2022-01-01']
# rides must be less than February 2022 -> year-month-day format
rides = rides[rides.pickup_datetime < '2022-02-01']
# check the describe function again
rides.pickup_datetime.describe()

### Save the Transformed Data to Subfolder

After our data is validated we should save it to a subfolder in our data folder. The **rides** dataframe can be saved as a parquet file using **.to_parquet** just like saving a csv file using **.to_csv**.

# save transformed file to another folder
rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
