==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

# read our file in 
rides = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')

rides.head()

We want to create a new column that rounds to the previous hour since we will be working on an hourly basis. To do so, we can perform actions on datetime objects by using **.dt**. Once we use that, we can use **.floor** which rounds down and specify 'h' because we want to round the hour. All this is doing is taking the pickup_datetime and making a new column called **pickup_hour** that rounds the time down to the nearest hour.


# add column of rounded hour
rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('h')

# verify 
rides

How many rides per location per hour? To get this information we need to use **.groupby**. We want to group our data so that we can see how many rides at each location happen per hour. Each row in our dataframe is a single ride that has the location ID and pickup hour included. This means there can be multiple rides to the same location per-hour. This information is beneficial to us because we can see how much demand we need for a given location in a given hour. 

After the data is grouped such that the pickup_location and pickup_hour have the same values, the **.size()** is used to create a new series that calculates the size, or number of occurences, of each group. Since this creates a series we use **.reset_index()** to convert it back to a data frame.

*be sure to use brackets [] in the groupby function since we are grouping by more than one column we use a list of columns.*

# create a column to count the number of rides per location per hour
agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()

# view new data frame
agg_rides

# rename the new column
agg_rides.rename(
    columns = {
    0: 'rides'
}, inplace = True)

agg_rides.head()

Next, we will make sure that each pickup hour is included in our data. For a large city like New York City, it is very unlikely that there are 0 taxi rides in a given hour but we still want to make sure that we are not excluding any hours from our data. If there are no taxi rides for a given location in a given hour we will impute a 0 so that we have no missing rows. This is done with a function and a for loop. Since for loops are very slow in Python, **tqdm** is imported which gives us a progress bar on the for loop.

from tqdm import tqdm

# a dataframe is given to this function that returns a dataframe
def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    # capture each location id using the .unique() function
    location_ids = agg_rides['pickup_location_id'].unique()

    # range of dates by hour this range has the smallest hour to the largest hour and a frequency based on hours.
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(),
        agg_rides['pickup_hour'].max(), 
        freq='h')
    # empty dataframe
    output = pd.DataFrame()

    # loop through each unique location using tqdm to get a progress bar
    for location_id in tqdm(location_ids):

        # keep only rides for this 'location_id' and return a dataframe with only pickup_hour and rides columns
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]
            
        # quick way to add missing dates with 0 in a Series
        # taken from https://stackoverflow.com/a/19324591
        # set index to the pickup hour 
        agg_rides_i.set_index('pickup_hour', inplace=True)
        # make it datetime index
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        # should have the full range of hours and fill empty with 0 
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        
        # add back `location_id` columns since we did not have them in the for loop
        agg_rides_i['pickup_location_id'] = location_id

        # fill the empty dataframe we initiated
        output = pd.concat([output, agg_rides_i])
    
    # our index column is the pickup_hour - change the name
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    
    return output

# call the function
agg_rides_all_slots = add_missing_slots(agg_rides)

### Plotting the Time Series Data

Before we plot the time series data we will import **Optional, List** from **typing**. This just makes our code more readable when creating functions.

from typing import Optional, List

import plotly.express as px

# function to plot the data using plotly express
# the locations are optional and will be a list if given
def plot_rides(rides: pd.DataFrame, locations: Optional[List[int]] = None):
    '''Plot the data'''
    # subset the data down to only rows where the pickup_location_id equals the list of locations given
    # if locations is not provided it is just the rides dataframe given
    rides_to_plot = rides[rides['pickup_location_id'].isin(locations)] if locations else rides

    # plot it 
    fig = px.line(
        rides_to_plot,
        x = 'pickup_hour',
        y = 'rides',
        color = 'pickup_location_id',
        template = 'none',
    )
     
    # show the plot
    fig.show()

# reminder of what rides is if no location is given
agg_rides.head()

When we plot the data for Central Park, there is obvious seasonality happening in the data.

# plot the data for central park
plot_rides(agg_rides_all_slots, locations=[43])

# save the aggregated data with no missing rows to parquet
agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
