==> notebooks/03_transform_ts_data_into_features_and_targets.ipynb <==
### Create Training Data from Time Series

The next step is to create our training data from the time series data of aggregated rides. The features in this data will be the previous rides per hour and the target will be how many rides in the hour we want to predict. To do this, we will use the slice and slide method. For example, if we want to predict the number of rides for the first hour of today, the features will be the previous 24 hours and the target will be the next hour. In terms of our data, this means the features will be indices 0, 24 from the data frame and the target will be the 25th index. For the next hour, we shift the bottom index up by one so the index wil be (1, 25, 26) with features being rows 1-25 and the target variable being row 26.

This tabular data has N + 1 columns where N is the number of features and the last column is the target variable. Remember, in Python, if we want to use the 12th hour as the target variable so that we use rows up to but not including the 12th row we would use range(0, 12) because 12 is not included in the range function.

import pandas as pd
import numpy as np

# import our aggregated data
ts_data = pd.read_parquet('../data/transformed/ts_data_2022_01.parquet')


To illustrate, here is an example that shows what the features will be if we are trying to predict the third hour. We will use two features, the previous two hours (11, 15), to predict the third unknown hour rides (26).

# example to predict the 26 rides in the third hour 
ts_data.head(n = 3)

# features to predict row 3
pred_3 = ts_data.iloc[0:2, 1].to_numpy()

# target variable
target_3 = ts_data.iloc[2, 1]

print(f'Features: {pred_3}\nTarget: {target_3}')


# ts data
ts_data.head()

# create central park data frame - central park is pickup location 43
ts_data_cp = ts_data.loc[ts_data['pickup_location_id'] == 43].reset_index(drop = True)

# central park time series df
ts_data_cp.head()

To create the features and targets we will need the tuple of indicies corresponding to the starting index, ending index, followed by the target variable. Since the length of data will always be 1 more than the highest index, the **stop_position** is len(data) - 1. 

# cutoff indices function
def get_cutoff_indices(data = pd.DataFrame, n_features = int, step_size = int) -> list:
    stop_position = len(data) - 1

    # start at 0th index - first feature
    subseq_first_idx = 0
    # last feature before target var
    subseq_mid_idx = n_features
    # this is the target
    subseq_last_idx = n_features + 1
    # empty list
    indices = []

    # continue this process for all data filling the empty list
    while subseq_last_idx <= stop_position:
            
            # add the three indices to our list
            indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))
            
            # our step size will be one so we add one to each
            subseq_first_idx += step_size
            subseq_mid_idx += step_size
            subseq_last_idx += step_size

    return indices


As we see below, the **get_cutoff_indices** function has generated training data for us by using 24 features for one target. The way to read this is that we use indices 0-24 as the features and 24 is not included. Then we use indices 24-25 for the target variable and 25 is not included so it is just index 24.

As a reminder, the indicies (0, 24, 25) will mean 0-23 for features since the last is not included.

# generate a sample where we use a full day as features
n_features = 24
step_size = 1

indices = get_cutoff_indices(
    ts_data_cp,
    n_features,
    step_size
)
indices[:5]

# we can see here that indices 0-23 account for one day (it is 24 hours)
ts_data_cp[0:28]

### Create X and y 

# view this to help visualize the following code
indices[0:2]

import numpy as np

n_features = 24
step_size = 1

# rows are the length of all sliced indices (list created above) and columns are the numbers of features
n_examples = len(indices)

# x is the first two numbers in our indices and y is just the length of the indices list
x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32) 
y = np.ndarray(shape=(n_examples), dtype=np.float32)
pickup_hours = []

# i is the index and idx is the element. The element is a tuple so idx[0] is the first number of the tuple
for i, idx in enumerate(indices):

    # x is assigned the rides value at the indices the loop is currently at 
    x[i, :] = ts_data_cp.iloc[idx[0]:idx[1]]['rides'].values

    # idx[1] is the target variable in the tuple since iloc is exclusive to ending index (24th index is target)
    y[i] = ts_data_cp.iloc[idx[1]:idx[2]]['rides'].values       

    # keep trach of pickup hours  
    pickup_hours.append(ts_data_cp.iloc[idx[1]]['pickup_hour'])

# view the creation of x
print(f'{x.shape=}')
print(f'{x=}')
print(f'{pickup_hours[:5]=}')


# change to dataframe 
features_one_location = pd.DataFrame(
    x, 
    # range is 24 but we want to iterate backwards so use reversed
    columns = [f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))]
)

# data frame of features
features_one_location

# do the same for target variables 
targets_one_location = pd.DataFrame(y, 
                                    columns = [f'target_rides_next_hour'])

# df of targets
targets_one_location

from tqdm import tqdm

def transform_ts_data_into_features_and_target(
    ts_data: pd.DataFrame,
    input_seq_len: int,
    step_size: int
) -> pd.DataFrame:
    """
    Slices and transposes data from time-series format into a (features, target)
    format that we can use to train Supervised ML models
    """
    # make sure the columns match exactly
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    # only need to do this once per location id
    location_ids = ts_data['pickup_location_id'].unique()

    # set up empty dataframes
    features = pd.DataFrame()
    targets = pd.DataFrame()
    
    # for each location in our time series data.....
    for location_id in tqdm(location_ids):
        
        # keep only ts data for this location_id
        ts_data_one_location = ts_data.loc[ts_data['pickup_location_id'] == location_id, ['pickup_hour', 'rides']]

        # indices to split dataframe rows
        indices = get_cutoff_indices(
            ts_data_one_location,
            # number of features
            input_seq_len,
            step_size
        )

        # slice and transpose data into numpy arrays for features and targets
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, input_seq_len), dtype=np.float32)
        y = np.ndarray(shape=(n_examples), dtype=np.float32)
        pickup_hours = []

        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values[0]
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

        # numpy -> pandas
        features_one_location = pd.DataFrame(
            x,
            columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(input_seq_len))]
        )
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        # numpy -> pandas
        targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])

        # concatenate results
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets['target_rides_next_hour']

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*7*1, # number of features in one week
    step_size=24,         # speed up the loop - 24 hour jump
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

features.head()
