==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
# Transform Raw Data into Features and Targets

### Step 1: Fetch/Validate the Data

Since we want to import another function from our data.py script but we already imported from src.data, Jupyter will have an **ImportError** because it chaches imports. This means it will revert back to the first import of src.data. To fix this, we use a magic command. This will tell Jupyter to reimport the model everytime you use an import statement.

%reload_ext autoreload
%autoreload 2

from src.data import load_raw_data

# load rides from 2022 for all months
rides = load_raw_data(year = 2022)
rides.head()

### Step 2: Transform the Data to Time Series Data

First we will have to transform the data into time series data using grouping and aggregation. We want to count the number of rides per-hour, per-pickup location. When we do this, we have to make sure we are imputing 0s where no rides take place. This function uses processes that we did in the 02_transform_raw_data_into_ts_data.ipynb notebook. 

from src.data import transform_raw_data_into_ts_data
# transform the raw data into time series data
ts_data = transform_raw_data_into_ts_data(rides)
ts_data

Just as before, the next step after transforming the data into time series data will be to transform that data into features and targets (tabular data transformation).

# import the function
from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data, 
    input_seq_len = 24*28*1,   # one month of features
    step_size = 24
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

features.pickup_hour[:5]

Now we follow a similar process as before when we saved the data to a parquet file. This time, we can use the **TRANSFORMED_DATA_DIR** variable that we created in the paths.py script. It is basically just the path that we want to store our data.

# save this data
tabular_data = features
tabular_data['target_rides_next_hour'] = targets

# save to transformed data path that we created in src data
from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

tabular_data
