==> notebooks/05_visualize_training_data.ipynb <==
## Visualize Training Data

In this notebook, we will be visualizing the training data. In our function to create the time series data, we made each month 28 days which is popular in data circles. Each row of data is a training sample. For example, the first row shows a full month of data. We will create a function to plot this data and store it in a separate python file called **plot.py**. 

The function will receive a dataframe of the features and a series for the targets with the option to also receive the predictions. We have not made any predictions at this point so we will not be providing any.

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

# read in our data from the TRANSFORMED_DATA_DIR
features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_targets

features = features_and_targets.drop(columns=['target_rides_next_hour'])
targets = features_and_targets['target_rides_next_hour']

from src.plot import plot_one_sample

plot_one_sample(
    example_id=1011,
    features=features,
    targets=targets,
)

The next row of data will start on January 2nd since we have a step size of 24 hours. It will then display 28 days of monthly data and the target variable will be exactly 1 day from the first target variable data point. This means this row will have the same amount of data points that the previous one has but it will start 1 day later and end 1 day later. Both will have target variable pickup times of 00:00:00 since our step size is 24 hours. 

# plot the next row
plot_one_sample(
    example_id=1012,
    features=features,
    targets=targets,
)
