==> notebooks/06_baseline_model.ipynb <==
## Create a Baseline Model

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

#### Split into Training and Testing

Now we will split the data into training and testing. We will not use the popular train_test_split from sklearn because we have created our own that will not split the data randomly but will instead split by a specified **cutoff_date**.

from src.data_split import train_test_split
from datetime import datetime

# cutoff_date must be datetime()
X_train, y_train, X_test, y_test = train_test_split(df,
                                                    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
                                                    target_column_name= 'target_rides_next_hour'
                                                    )

print(f'{X_train.shape}')
print(f'{y_train.shape}')
print(f'{X_test.shape}')
print(f'{y_test.shape}')


#### First Baseline Model - Predict Using Last Hour

Since this is our first baseline model we are going to make this model very simple. We are going to predict the taxi demand to be the demand that it was in the previous hour. We do this by making a class. Normally most of my code uses functions but we want this to follow the same structure of our next models which will be classes with fit and predict options.

This class will have fit and predict methods within it. The fit method will actually do nothing since our model is so simple that it will not have to be trained. The predict method will just have to return the **rides_previous_1_hour** column for each value in our **X_test** data. 

import numpy as np

class BaselineModelPreviousHour:
    '''class for baseline 1 model'''
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        # this model is very simple and does nothing 
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        # just return the rides from the previous hour
        return X_test[f'rides_previous_1_hour']




# instantiate the class 
model = BaselineModelPreviousHour()
# make predictions - just the rides_previous_1_hour column of that row
predictions = model.predict(X_test)
# view
predictions

##### Error Metric - MAE

Average absolute difference between predicted values and actual values. This means on average our demand is off by 6.05 units, or rides. The mean number of rides is roughly 12 so this is a very bad model. The number 6 may not seem like a lot but when the average number of rides is only 12 it means we are off by a lot. 

y_test.mean()

from sklearn.metrics import mean_absolute_error

test_mea = mean_absolute_error(y_test, predictions)
print(f'{test_mea=:.4f}')

#### Second Baseline Model - Use Last Week's Demand

This model will use the value of last week's demand for this week's demand. This model has an MAE of 3.68 which is much better than the original model. We thought this would be the case when we viewed the central park time series because there was weekly seasonality present in the data. This goes to show that EDA is important.

class BaselineModelPreviousWeek:
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        '''remember rides_previous_#_hour works backwards from todays demand.'''
        return X_test[f'rides_previous_{7*24}_hour']
    
    

# create object from second baseline class
model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)


# check the mae - the = inside the curly brackets tells python to include the variable name in the print statement.
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

Another option would be to create a model based on the previous days demand but this model will not be as good as last week because weekends will have more demand than weekdays so going based on the previous week is better. 

class Baseline24:
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        '''remember rides_previous_#_hour works backwards from todays demand.'''
        return X_test[f'rides_previous_{1*24}_hour']

# create object from second baseline class
model = Baseline24()
predictions = model.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

#### Third Baseline Model - Last 4 Week's Average

For this baseline model, we will build on our weekly seasonality intuition by getting the average over the last 4 weeks. This model performs the best out of our baseline models.

class BaselineModelPrevious4Weeks():
    def fit(self, X_train: pd.DataFrame, y_train:pd.Series):
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        return (X_test[f'rides_previous_{7*24}_hour'] + \
        X_test[f'rides_previous_{14*24}_hour'] + \
        X_test[f'rides_previous_{21*24}_hour'] + \
        X_test[f'rides_previous_{28*24}_hour']) / 4

model = BaselineModelPrevious4Weeks()
predictions = model.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
