==> notebooks/07_xgboost_model.ipynb <==
## XGBOOST Model

# import pandas
import pandas as pd

# import transformed data dir
from src.paths import TRANSFORMED_DATA_DIR

# import and load the data
df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

# train test split using june 1 2022
from datetime import datetime

from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(df,
                                                    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
                                                    target_column_name='target_rides_next_hour')

# view shape of each
print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')



import xgboost as xgb

Input features to make predictions will be the past rides. This means we will not use the pickup_hour column at this point.

# make dataframe that excludes pickup_hour
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

X_train_only_numeric.shape

# instantiate the class
model = xgb.XGBRegressor()
model.fit(X_train_only_numeric, y_train)

# generate predictions on only past ride columns
X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)

# print predictions 
predictions

# get mae of our model
from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)

# view mae to 4th decimal
print(f'{test_mae=:.4f}')

The mae of 2.7082 beats all baseline models. However, XGBOOST works slowly on the type of dataset that we have (many features), so we will create a LightGBM model to make it more efficient memorywise.
