==> notebooks/09_lightgbm_with_feature_engineering.ipynb <==
## LightGBM with Feature Engineering

# load the data
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR /'tabular_data.parquet')

df.head()

The first 3 drop off locations have almost no rides so I want to remove them from the data. 

df.head()

# train and test sets
from src.data_split import train_test_split
from datetime import datetime

# create training and testing sets
X_train, y_train, X_test, y_test = train_test_split(df, cutoff_date=datetime(2022, 6, 1, 0, 0, 0), 
                                                    target_column_name= 'target_rides_next_hour')

# view shapes
print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')


We are going to use Feature Engineering and Scikit-Learn pipelines for our data. The first feature engineering step is to add a feature/column to our data for the previous four week average rides. This was also done as a baseline model. Since we want this to be part of the pipeline, we need to use **FunctionTransformer** if we want it to be a step included in our pipeline.

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['average_rides_last_4_weeks'] = 0.25*(
        X[f'rides_previous_{7*24}_hour'] + \
        X[f'rides_previous_{2*7*24}_hour'] + \
        X[f'rides_previous_{3*7*24}_hour'] + \
        X[f'rides_previous_{4*7*24}_hour']
    )
    return X

from sklearn.preprocessing import FunctionTransformer

add_feature_previous_4_average = FunctionTransformer(average_rides_last_4_weeks, validate = False)

Now we will use the **Custom Transformer** capability from sklearn to create a column for the hour of the day and the day of the week.

*note: the column for the hour of the day will always be hour 0 because we sliced every 24 hours. The reason we are including it in our analysis is incase we want to change the slicing in the future to predict every hour.*

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeatures(BaseEstimator, TransformerMixin):
    def fit(self, X, y = None):  # we have no parameters so nothing has to be done
        return self
    def transform(self, X, y = None):
        # make a copy of X since we are dropping a row
        X_ = X.copy()

        # hour of day and day of week
        X_['hour'] = X_['pickup_hour'].dt.hour
        X_['day'] = X_['pickup_hour'].dt.dayofweek

        # using lightGBM so we have to drop the datetime column
        return X_.drop(columns = ['pickup_hour'])

# implement previous two steps
add_feature_previous_4_average.fit_transform(X_train)
add_temporal_features = TemporalFeatures()
add_temporal_features.fit_transform(X_train)

# put all steps together with a pipeline
from sklearn.pipeline import Pipeline
import lightgbm as lgb

pipeline = Pipeline([
    ('prev_4', add_feature_previous_4_average),
    ('temporal_features', TemporalFeatures()),
    ('lgb', lgb.LGBMRegressor()),
])

# fit transform
pipeline.fit(X_train, y_train)

from sklearn.metrics import mean_absolute_error
# make predictions 
predictions = pipeline.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)

# print the results
print(f'{test_mae=:.4f}')
