==> notebooks/11_backfill_feature_store.ipynb <==
import hopsworks

HOPSWORKS_PROJECT_NAME = 'taxi_demand_rs'

import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR

# load key-value pairs from .env file located in the parent directory
load_dotenv(PARENT_DIR / '.env')

HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']

I want to only use data from 2023 onward since I was getting a message about hitting a row limit with the data. We will see if this causes any issues to the model.

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2023
to_year = datetime.now().year
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # append rows
    rides = pd.concat([rides, rides_one_year])

Now we have all rides from the start of 2023 until the end of July 2024. This is because the site we are getting data from has not uploaded any data past this date.

rides

We have 61 million rides now.

Here is where things get just a little bit tricky. This is where all future data will become synthetic data that is actually just data from a year ago. The first step is to add in the August 2024 data which will be the August 2023 data.

# create to and from dates starting with August
august = pd.to_datetime("2024-08-01 00:00:00+00:00")

# til date is Sep 11th
sep = pd.to_datetime('2024-09-11 20:00:00+00:00')

print(f'Fetching synthetic data from {august} until {sep}')


from src.data import load_raw_data
from datetime import datetime, timedelta
import pandas as pd
import pytz

def fetch_ride_events_from_data_warehouse(
    from_date: datetime,
    to_date: datetime
) -> pd.DataFrame:
    """
    This function is used to simulate production data by sampling historical data
    from 52 weeks ago (i.e. 1 year)
    """
    # Ensure from_date and to_date are timezone-aware
    from_date = from_date.replace(tzinfo=pytz.UTC)
    to_date = to_date.replace(tzinfo=pytz.UTC)

    from_date_ = pd.Timestamp(from_date - timedelta(days=7*52))
    to_date_ = pd.Timestamp(to_date - timedelta(days=7*52))
    print(f'Fetching ride events from {from_date} to {to_date}')

    if (from_date_.year == to_date_.year) and (from_date_.month == to_date_.month):
        # download 1 file of data only
        rides = load_raw_data(year=from_date_.year, months=from_date_.month)
        rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime'], utc=True)
        rides = rides[rides.pickup_datetime >= from_date_]
        rides = rides[rides.pickup_datetime < to_date_]
    else:
        # download 2 files from website
        rides = load_raw_data(year=from_date_.year, months=from_date_.month)
        rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime'], utc=True)
        rides = rides[rides.pickup_datetime >= from_date_]

        rides_2 = load_raw_data(year=to_date_.year, months=to_date_.month)
        rides_2['pickup_datetime'] = pd.to_datetime(rides_2['pickup_datetime'], utc=True)
        rides_2 = rides_2[rides_2.pickup_datetime < to_date_]

        rides = pd.concat([rides, rides_2])

    # shift the pickup_datetime back 1 year ahead, to simulate production data
    # using its 7*52-days-ago value
    rides['pickup_datetime'] = rides['pickup_datetime'] + timedelta(days=7*52)

    rides.sort_values(by=['pickup_location_id', 'pickup_datetime'], inplace=True)

    return rides

rides_synthetic = fetch_ride_events_from_data_warehouse(august, sep)
rides_synthetic

len(rides)

rides

# Convert pickup_datetime to datetime and make it timezone-aware
rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime']).dt.tz_localize('UTC')

rides

rides_synthetic

# add this to rides data
rides = pd.concat([rides, rides_synthetic])

rides

print(f'{len(rides)=:,}')

rides.info()

rides.sort_values(by=['pickup_datetime', 'pickup_location_id'], inplace=True)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

Now we have the rides per hour for each pickup_location.

ts_data

len(ts_data.loc[ts_data['pickup_location_id'] == 265])

# Convert string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# Add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype('int64') // 10**6

This operation added a **pickup_ts** column to help hopsworks with the event time primary key.

ts_data

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

Now we will write data to the feature group within our feature store.

len(ts_data)

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_ts'],
    event_time='pickup_ts',
)

# add this data to feature group version 2
feature_group.insert(ts_data, write_options={"wait_for_job": True})


df = feature_group.read()

len(df)

df.tail()
