==> notebooks/14_inference_pipeline.ipynb <==
## Inference Pipeline

%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pytz
import pandas as pd

current_date = pd.to_datetime(datetime.now(pytz.utc)).floor('H')
print(f'{current_date=}')

The following cell will fetch a batch of raw features just as we did when we made our **frontend.py** app. This downloads rides from the past 28 days. Just as in our frontend app, this function will not work unless the most recent data is available to us. Using the GitHub actions, these features will not be available for the first 15-25 minutes of every hour or so. This is why this notebook will not be run unless our feature_pipeline is completed.

from src.inference import load_batch_of_features_from_store

features = load_batch_of_features_from_store(current_date)

##### Load the Model

Now, we will connect to the project and model registry, download the model, and load it to a local file and get predictions from the model using the new features.

The **get_model_predictions** function returns a dataframe with the pickup_location_id and predicted demand rounded to the nearest integer. The predictions are the number of rides for the current hour since the features are the previous hours over the last 28 days up until an hour ago.

from src.inference import(
    load_model_from_registry,
    get_model_predictions
)

model = load_model_from_registry()
predictions = get_model_predictions(model=model, features=features)

predictions

We are creating predictions that will be our feature group. The next steps are very similar to what we did when we did our feature_pipeline. We need to create a pickup_hour and pickup_ts column (hopsworks does not like datetime), then create the feature group and write the data to it.

predictions['pickup_hour'] = current_date
predictions['pickup_ts'] = predictions['pickup_hour'].astype('int64') // 10**6

predictions

from src.inference import get_feature_store
import src.config as config

feature_store = get_feature_store()

feature_group = feature_store.get_or_create_feature_group(
    name = config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version= 1,
    description= 'Predictions generated by model',
    primary_key= ['pickup_location_id', 'pickup_ts'],
    event_time= 'pickup_ts'
)





# write the data to the new feature group
feature_group.insert(predictions, write_options={'wait_for_job' : True})
